+++ components/FileFormats.vue
@@ -1,6 +1,20 @@
 <template>
 	<div class="vue-component file-formats">
-		<SearchableList :data="fileFormats" summaryKey="title" keywordsKey="gis_data_types" :showSummaryOnExpand="false" :externalSearchTerm="searchTerm" :externalHideDeprecated="hideDeprecated" :deprecatedFilter="deprecatedFilter" :sort="sort" :offerDetails="offerDetails" :heading="heading" :collapsed="collapsed" @detailsToggled="detailsToggled">
+		<SearchableList
+			:data="fileFormats"
+			summaryKey="title"
+			keywordsKey="gis_data_types"
+			:showSummaryOnExpand="false"
+			:externalSearchTerm="searchTerm"
+			:externalHideDeprecated="hideDeprecated"
+			:deprecatedFilter="deprecatedFilter"
+			:externalHideExperimental="hideExperimental"
+			:experimentalFilter="experimentalFilter"
+			:sort="sort"
+			:offerDetails="offerDetails"
+			:heading="heading"
+			:collapsed="collapsed"
+			@detailsToggled="detailsToggled">
 			<template #heading="scope"><slot name="heading" v-bind="scope" /></template>
 			<template v-if="missing" #content-start><FederationMissingNotice :missing="missing" :federation="federation" /></template>
 			<template #summary="slot">
@@ -79,6 +93,14 @@ export default {
 			default: false
 		},
 		deprecatedFilter: {
+			type: Boolean,
+			default: false
+		},
+		hideExperimental: {
+			type: Boolean,
+			default: false
+		},
+		experimentalFilter: {
 			type: Boolean,
 			default: false
 		},